Make the range fixer idempotent
Along the way, move the fixer over to `lib2to3.fixes.fix_xrange` to
piggyback on all the work done there.

Closes #52 .
