+++ libmodernize/fixes/__init__.py
@@ -33,9 +33,9 @@ six_fix_names = set([
     'libmodernize.fixes.fix_filter',
     'libmodernize.fixes.fix_map',
     'libmodernize.fixes.fix_metaclass',
-    'libmodernize.fixes.fix_range',
     'libmodernize.fixes.fix_raise_six',
     'libmodernize.fixes.fix_unicode',
     'libmodernize.fixes.fix_unicode_type',
+    'libmodernize.fixes.fix_xrange_six',
     'libmodernize.fixes.fix_zip',
     ])
